Update - Remove update until someone fix it 100%, since u can deploy easily with  workflow. - Remove config(permanently) check heroku guide changes - Change torrent timeout for aria/qbit u can edit no. of sec by ur self Signed-off-by: anas <e.anastayyar@gmail.com>
