+++ alive.py
@@ -5,14 +5,6 @@
 import time
 import requests
 import os
-import subprocess
-from dotenv import load_dotenv
-
-CONFIG_FILE_URL = os.environ.get('CONFIG_FILE_URL', None)
-if CONFIG_FILE_URL is not None:
-    out = subprocess.run(["wget", "-q", "-O", "config.env", CONFIG_FILE_URL])
-
-load_dotenv('config.env')
 
 BASE_URL = os.environ.get('BASE_URL_OF_BOT', None)
 if len(BASE_URL) == 0:
