+++ aria.sh
@@ -1,7 +1,7 @@
 tracker_list=$(curl -Ns https://raw.githubusercontent.com/XIU2/TrackersListCollection/master/all.txt https://ngosang.github.io/trackerslist/trackers_all_http.txt https://newtrackon.com/api/all https://raw.githubusercontent.com/DeSireFire/animeTrackerList/master/AT_all.txt https://raw.githubusercontent.com/hezhijie0327/Trackerslist/main/trackerslist_tracker.txt https://raw.githubusercontent.com/hezhijie0327/Trackerslist/main/trackerslist_exclude.txt | awk '$0' | tr '\n\n' ',')
 aria2c --enable-rpc --check-certificate=false \
    --max-connection-per-server=10 --rpc-max-request-size=1024M --bt-max-peers=0 \
-   --bt-stop-timeout=1200 --min-split-size=10M --follow-torrent=mem --split=10 \
+   --bt-stop-timeout=0 --min-split-size=10M --follow-torrent=mem --split=10 \
    --daemon=true --allow-overwrite=true --max-overall-download-limit=0 --bt-tracker="[$tracker_list]"\
    --max-overall-upload-limit=1K --max-concurrent-downloads=15 --continue=true \
    --peer-id-prefix=-qB4360- --user-agent=qBittorrent/4.3.6 --peer-agent=qBittorrent/4.3.6 \
