+++ bot/__init__.py
@@ -33,7 +33,15 @@ logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s
 
 LOGGER = logging.getLogger(__name__)
 
+CONFIG_FILE_URL = os.environ.get('CONFIG_FILE_URL', None)
+if CONFIG_FILE_URL is not None:
+    out = subprocess.run(["wget", "-q", "-O", "config.env", CONFIG_FILE_URL])
+    if out.returncode != 0:
+        logging.error(out)
+
 load_dotenv('config.env')
+
+alive = subprocess.Popen(["python3", "alive.py"])
 
 Interval = []
 
@@ -129,8 +137,6 @@ try:
     AUTO_DELETE_MESSAGE_DURATION = int(getConfig('AUTO_DELETE_MESSAGE_DURATION'))
     TELEGRAM_API = getConfig('TELEGRAM_API')
     TELEGRAM_HASH = getConfig('TELEGRAM_HASH')
-    UPSTREAM_REPO = getConfig('UPSTREAM_REPO')
-    UPSTREAM_BRANCH = getConfig('UPSTREAM_BRANCH')
 except KeyError as e:
     LOGGER.error("One or more env variables missing! Exiting now")
     exit(1)
@@ -196,13 +202,12 @@ except KeyError:
     MEGA_PASSWORD = None
 try:
     HEROKU_API_KEY = getConfig('HEROKU_API_KEY')
-except KeyError:
-    logging.warning('HEROKU API KEY not provided!')
-    HEROKU_API_KEY = None
-try:
     HEROKU_APP_NAME = getConfig('HEROKU_APP_NAME')
+    if len(HEROKU_API_KEY) == 0 or len(HEROKU_APP_NAME) == 0:
+        HEROKU_API_KEY = None
+        HEROKU_APP_NAME = None
 except KeyError:
-    logging.warning('HEROKU APP NAME not provided!')
+    HEROKU_API_KEY = None
     HEROKU_APP_NAME = None
 try:
     UPTOBOX_TOKEN = getConfig('UPTOBOX_TOKEN')
@@ -319,14 +324,14 @@ try:
 except KeyError:
     SHORTENER = None
     SHORTENER_API = None
-
-IGNORE_PENDING_REQUESTS = False
 try:
-    if getConfig("IGNORE_PENDING_REQUESTS").lower() == "true":
+    IGNORE_PENDING_REQUESTS = getConfig("IGNORE_PENDING_REQUESTS")
+    if IGNORE_PENDING_REQUESTS.lower() == 'true':
         IGNORE_PENDING_REQUESTS = True
+    else:
+        IGNORE_PENDING_REQUESTS = False
 except KeyError:
-    pass
-
+    IGNORE_PENDING_REQUESTS = False
 try:
     BASE_URL = getConfig('BASE_URL_OF_BOT')
     if len(BASE_URL) == 0:
@@ -334,7 +339,6 @@ try:
 except KeyError:
     logging.warning('BASE_URL_OF_BOT not provided!')
     BASE_URL = None
-
 try:
     IS_VPS = getConfig('IS_VPS')
     if IS_VPS.lower() == 'true':
@@ -343,7 +347,6 @@ try:
         IS_VPS = False
 except KeyError:
     IS_VPS = False
-
 try:
     SERVER_PORT = getConfig('SERVER_PORT')
     if len(SERVER_PORT) == 0:
@@ -351,7 +354,6 @@ try:
 except KeyError:
     logging.warning('SERVER_PORT not provided!')
     SERVER_PORT = None
-
 try:
     TOKEN_PICKLE_URL = getConfig('TOKEN_PICKLE_URL')
     if len(TOKEN_PICKLE_URL) == 0:
@@ -362,7 +364,6 @@ try:
             logging.error(out)
 except KeyError:
     TOKEN_PICKLE_URL = None
-
 try:
     ACCOUNTS_ZIP_URL = getConfig('ACCOUNTS_ZIP_URL')
     if len(ACCOUNTS_ZIP_URL) == 0:
