+++ bot/helper/__init__.py
@@ -28,27 +28,3 @@ def check_heroku(func):
                 await func(client, message, heroku_app)
 
     return heroku_cli
-
-# Preparing For Update Bot
-# Implement by https://github.com/jusidama18 and Based on this https://github.com/DevsExpo/FridayUserbot/blob/master/plugins/updater.py
-
-def fetch_heroku_git_url(api_key, app_name):
-    if not api_key:
-        return None
-    if not app_name:
-        return None
-    heroku = heroku3.from_key(api_key)
-    try:
-        heroku_applications = heroku.apps()
-    except:
-        return None
-    heroku_app = None
-    for app in heroku_applications:
-        if app.name == app_name:
-            heroku_app = app
-            break
-    if not heroku_app:
-        return None
-    return heroku_app.git_url.replace("https://", "https://api:" + api_key + "@")
-
-HEROKU_URL = fetch_heroku_git_url(HEROKU_API_KEY, HEROKU_APP_NAME)
