+++ config_sample.env
@@ -11,8 +11,6 @@ AUTO_DELETE_MESSAGE_DURATION = 20
 IS_TEAM_DRIVE = ""
 TELEGRAM_API = 
 TELEGRAM_HASH = ""
-UPSTREAM_REPO = "https://github.com/SlamDevs/slam-mirrorbot"
-UPSTREAM_BRANCH = "master"
 # OPTIONAL CONFIG
 DATABASE_URL = ""
 AUTHORIZED_CHATS = ""  # Split by space
@@ -31,7 +29,7 @@ STOP_DUPLICATE = ""
 SHORTENER = ""
 SHORTENER_API = ""
 # qBittorrent
-IS_VPS = ""
+IS_VPS = ""  # Don't set this to true even if vps, unless facing error with web server
 SERVER_PORT = "80"  # Only for VPS
 BASE_URL_OF_BOT = ""  # Required for Heroku
 # If you want to use Credentials externally from Index Links, fill these vars with the direct links
