+++ heroku-guide.md
@@ -20,13 +20,20 @@
 	* `HEROKU_EMAIL` Heroku Account Email Id in which the above app will be deployed
 	* `HEROKU_API_KEY` Your Heroku API key, get it from https://dashboard.heroku.com/account
 	* `HEROKU_APP_NAME` Your Heroku app name, Name Must be unique
-	* `CONFIG_FILE_URL` Fill [This](https://raw.githubusercontent.com/Slam-Team/slam-mirrorbot/master/config_sample.env) in any text editor, and remove the second line. Go to https://gist.github.com and paste your config data from the previous step. Rename the file to config.env and upload it. Click on Raw, copy the link and that's your CONFIG_FILE_URL. Refer to below images for clarity. 
+	* `CONFIG_FILE_URL` Fill [This](https://raw.githubusercontent.com/Slam-Team/slam-mirrorbot/master/config_sample.env) in any text editor. Remove the _____REMOVE_THIS_LINE_____ line and fill the variables. Go to https://gist.github.com and paste your config data. Rename the file to config.env then create secret gist. Click on Raw, copy the link. This will be your CONFIG_FILE_URL. Refer to below images for clarity. 
 
 	![steps 1 to 3](https://telegra.ph/file/1d8fec16516a87ba9d1ac.jpg)
 
 	![step 4](https://telegra.ph/file/1491f99836cd694ea1195.jpg)
 
 	![step 5](https://telegra.ph/file/416a550f7ded579b63272.jpg)
+
+- Note: Remove commit id from raw link to be able to change variables without updating the CONFIG_FILE_URL in secrets.
+  should be in this form: https://gist.githubusercontent.com/username/gist-id/raw/config.env
+  - Before: https://gist.githubusercontent.com/anasty17/8cce4a4b4e7f4ea47e948b2d058e52ac/raw/19ba5ab5eb43016422193319f28bc3c7dfb60f25/config.env
+  - After: https://gist.githubusercontent.com/anasty17/8cce4a4b4e7f4ea47e948b2d058e52ac/raw/config.env
+
+  - You only need to restart your bot after editing config.env gist secret.
 
 4. After adding all the above Required Variables go to Github Actions tab in your repo
